refactor: 중복이름, Null이름 검증을 Cars에서 수행하도록 리팩토링
