+++ src/main/java/racingcar/controller/RacingCarController.java
@@ -9,7 +9,6 @@ import java.util.*;
 
 public class RacingCarController {
     private static final String DELIMITER = ",";
-    private static final int UPPER_LIMIT = 5;
     private static Scanner scanner;
     private Cars cars;
     private int turns;
@@ -18,49 +17,11 @@ public class RacingCarController {
         this.scanner = scanner;
     }
 
-    public static void checkNames(List<String> nameCandidates) {
-        checkOverlappedNames(nameCandidates);
-        checkNameLength(nameCandidates);
-        checkNull(nameCandidates);
-    }
-
-    private static void checkOverlappedNames(List<String> nameCandidates) {
-        Set<String> targetSet = new HashSet<>(nameCandidates);
-        if (nameCandidates.size() != targetSet.size()) {
-            RacingCarError.overlapped();
-        }
-    }
-
-    private static void checkNameLength(List<String> nameCandidates) {
-        checkUpperLimit(nameCandidates);
-        checkLowerLimit(nameCandidates);
-    }
-
-    private static void checkUpperLimit(List<String> nameCandidates) {
-        long nameFilters = nameCandidates.stream()
-                .mapToInt(carName -> carName.length())
-                .filter(carNameLength -> carNameLength > UPPER_LIMIT)
-                .count();
-        if (nameFilters != 0) {
-            RacingCarError.upperLength();
-        }
-    }
-
-    private static void checkLowerLimit(List<String> nameCandidates) {
-        long nameFilters = nameCandidates.stream()
-                .mapToInt(carName -> carName.length())
-                .filter(carNameLength -> carNameLength <= 0)
-                .count();
-        if (nameFilters != 0) {
-            RacingCarError.lowerLength();
-        }
-    }
-
-    private static void checkNull(List<String> nameCandidates) {
-        if (nameCandidates.size() == 0) {
-            RacingCarError.nullCar();
-        }
-    }
+//    public static void checkNames(List<String> nameCandidates) {
+//        checkOverlappedNames(nameCandidates);
+//        checkNameLength(nameCandidates);
+//        checkNull(nameCandidates);
+//    }
 
     public static void checkTurns(String turnCandidate) {
         int integerCandidate = checkInteger(turnCandidate);
@@ -82,14 +43,14 @@ public class RacingCarController {
         }
     }
 
-    public static void generateCars() {
-        RacingCarView.carListInput();
-        String carsInput = scanner.nextLine();
-        String[] parsedCarsInput = carsInput.split(DELIMITER);
-        List<String> carNameCandidates = Arrays.asList(parsedCarsInput);
-        checkNames(carNameCandidates);
-        this.cars = new Cars(carNameCandidates);
-    }
+//    public static void generateCars() {
+//        RacingCarView.carListInput();
+//        String carsInput = scanner.nextLine();
+//        String[] parsedCarsInput = carsInput.split(DELIMITER);
+//        List<String> carNameCandidates = Arrays.asList(parsedCarsInput);
+//        checkNames(carNameCandidates);
+//        this.cars = new Cars(carNameCandidates);
+//    }
 
     public void receiveTurns() {
         RacingCarView.turnNumberInput();
